AI_Study_App/backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=1.10.0
urllib3>=1.26.0
